refactor: prepare for other uses of the ast.File
